Functional Stock Predictor.
A Tkinter GUI interfaces with the user, simply enter the Symbol of your stock (not case dependent, string manipulation to take care of that), and the application returns about 10 charts showing different projections with different factors using PyPlot. Prediction method is an LSTM from Keras and Tensorflow. Project is still in development; Pyinstaller does not have an effective way of transforming Tensorflow into an application. This may be fixed in the future. For now, the program will run only in an IDE, or from shell execution. The program will also only work on Windows, due to dependability issues with specific modules.